styling readjustments; moved pagination code to its own file component
