+++ app/src/routes/_storefront/index.tsx
@@ -4,10 +4,7 @@ import { ChildCard, ChildCardData } from "@/components/storefront/ChildCard";
 import { Child, Family } from "../../../../common/src/types";
 import { useState } from "react";
 import { z } from "zod";
-import { Button } from "@/components/ui/button";
-import { 
-  ChevronDoubleLeftIcon, ChevronDoubleRightIcon 
-} from "@/components/icons";
+import { Pagination } from "@/components/storefront/Pagination";
 
 export const Route = createFileRoute("/_storefront/")({
   // STEP 1: define the search param for TanStack Router
@@ -19,80 +16,6 @@ export const Route = createFileRoute("/_storefront/")({
 });
 
 const familyColors = ["kfk-red", "kfk-brown", "kfk-green", "kfk-blue"];
-
-const getPaginationRange = (
-  pages:Array<number>,
-  currentPage: number,
-  totalPages: number,
-  maxButtons: number,
-  immediatePages: number
-) => {
-  let range = [];
-
-  if (pages.length <= maxButtons)
-    range = pages;
-  else if (currentPage <= immediatePages)
-    range = [...pages.slice(0, immediatePages + 3), -1, totalPages]; // -1 represents the ellipsis
-  else if (currentPage > totalPages - immediatePages)
-    range = [1, -1, ...pages.slice(totalPages - immediatePages - 3, totalPages)];
-  else
-    range = [1, -1, ...pages.slice(currentPage - 1 - immediatePages/2, currentPage + immediatePages/2), -1, totalPages];
-
-  return range;
-}
-
-// Pagination
-interface PaginationProp {
-  totalChildren: number,
-  childrenPerPage: number,
-  setCurrentPage: (page: number) => any,
-  currentPage: number,
-  MAX_BUTTONS: number,
-  IMMEDIATE_PAGES: number,
-}
-const Pagination = ( {totalChildren, childrenPerPage, setCurrentPage, currentPage, MAX_BUTTONS, IMMEDIATE_PAGES} : PaginationProp ) => {
-  const totalPages = Math.ceil(totalChildren/childrenPerPage)
-  let pages = [];
-
-  for (let i = 1; i <= totalPages; i++) {
-    pages.push(i);
-  }
-
-  const range = getPaginationRange(pages, currentPage, totalPages, MAX_BUTTONS, IMMEDIATE_PAGES);
-
-  return (
-    <div className="mx-auto mt-10 flex justify-center gap-2 text-xl">
-      {/* Left Arrow */}
-      <Button 
-        onClick={() => setCurrentPage(currentPage - 1)}
-        disabled={currentPage == 1}
-        className={`rounded-full pl-2 w-10 h-10 ${currentPage == 1 ? "text-gray-400" : "bg-transparent text-primary hover:bg-gray-100"} transition-all text-xl`}
-      >
-        <ChevronDoubleLeftIcon className="size-5"/>
-      </Button>
-      {/* Pages */}
-      {range.map((page) => {
-        return (
-          <Button 
-            onClick={() => setCurrentPage(page)}
-            disabled={page == -1}
-            className={`rounded-full text-primary w-10 h-10 ${page != -1 ? page == currentPage ? "bg-kfk-blue text-white" : "bg-transparent hover:bg-gray-100" : ""} transition-all text-xl`}
-          >
-            {page == -1 ? "..." : page}
-          </Button>
-        )
-      })}
-      {/* Right Arrow */}
-      <Button 
-        onClick={() => setCurrentPage(currentPage + 1)}
-        disabled={currentPage == Math.ceil(totalChildren/childrenPerPage)}
-        className={`rounded-full pl-3 w-10 h-10 ${currentPage == totalPages ? "text-gray-400" : "bg-transparent text-primary hover:bg-gray-100"} transition-all text-xl`}
-      >
-        <ChevronDoubleRightIcon className="size-5"/>
-      </Button>
-    </div>
-  )
-}
 
 // Mock families
 const mockFamilies: Family[] = [
@@ -210,7 +133,7 @@ const mockChildren: (ChildCardData & { familyId: string })[] =
   ;
 
 function App() {
-  const [childrenPerPage] = useState<number>(25);
+  const [childrenPerPage] = useState<number>(1);
   const [currentPage, setCurrentPage] = useState<number>(1);
  
   // STEP 2: read search/sort param
